requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
augly==1.0.0
build==1.2.2.post1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
einops==0.8.1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
invisible-watermark==0.2.0
iopath==0.1.10
Jinja2==3.1.6
kornia==0.7.3
kornia_rs==0.1.9
lazy_loader==0.4
lightning==2.3.3
lightning-utilities==0.11.9
lpips==0.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pathlib==1.0.1
pillow==10.4.0
pip-tools==7.4.1
portalocker==3.0.0
propcache==0.2.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytorch-lightning==2.4.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-image==0.21.0
scipy==1.10.1
shellingham==1.5.4
six==1.17.0
sympy==1.13.3
tifffile==2023.7.10
timm==1.0.17
tomli==2.2.1
torch==2.4.1
torchmetrics==1.5.2
torchvision==0.19.1
tqdm==4.67.1
triton==3.0.0
typer==0.16.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
yarl==1.15.2
zipp==3.20.2
